Add FrontPanel Client class to maintain connection to a FrontPanel over IP Server. This allows the possibility of passing the instance to other classes that need to communicate with the Server.
